+++ index.js
@@ -2,6 +2,7 @@ import { AppRegistry } from 'react-native'
 import BackgroundGeolocation from 'react-native-background-geolocation'
 import 'react-native-gesture-handler'
 import 'react-native-url-polyfill/auto'
+import 'text-encoding-polyfill'
 
 import App from './src/App'
 import { GeolocationTrackingHeadlessTask } from './src/app/domain/geolocation/tracking'
