+++ src/components/webviews/CryptoWebView/cryptoObservable/typedefs.js
@@ -6,7 +6,7 @@
  */
 
 /**
- * @typedef {('computePass'|'computePKCE')} CryptoMethodName
+ * @typedef {('computePass')} CryptoMethodName
  */
 
 /**
